:recycle: Refactor: shared/ui 컴포넌트 분리, 타입 설정, 폴더 정리
- index.ts에는 외부에서 사용할 것만 export로 정의하기
