+++ src/shared/ui/card/index.ts
@@ -1,0 +1,4 @@
+export { CardContainer } from "./CardContainer.tsx"
+export { CardHeader } from "./CardHeader.tsx"
+export { CardTitle } from "./CardTitle.tsx"
+export { CardContent } from "./CardContent.tsx"
