+++ src/shared/ui/common/index.ts
@@ -1,0 +1,3 @@
+export { Button } from "./Button.tsx"
+export { Input } from "./Input.tsx"
+export { Textarea } from "./Textarea.tsx"
