+++ src/shared/ui/dialog/DialogContainer.tsx
@@ -1,0 +1,3 @@
+import * as DialogPrimitive from "@radix-ui/react-dialog"
+
+export const DialogContainer = DialogPrimitive.Root
