+++ src/shared/ui/dialog/DialogPortal.tsx
@@ -1,0 +1,3 @@
+import * as DialogPrimitive from "@radix-ui/react-dialog"
+
+export const DialogPortal = DialogPrimitive.Portal
