+++ src/shared/ui/dialog/index.ts
@@ -1,0 +1,4 @@
+export { DialogContainer } from "./DialogContainer.tsx"
+export { DialogContent } from "./DialogContent.tsx"
+export { DialogHeader } from "./DialogHeader.tsx"
+export { DialogTitle } from "./DialogTitle.tsx"
