+++ src/shared/ui/select/SelectContainer.tsx
@@ -1,0 +1,3 @@
+import * as SelectPrimitive from "@radix-ui/react-select"
+
+export const SelectContainer = SelectPrimitive.Root
