+++ src/shared/ui/select/index.ts
@@ -1,0 +1,6 @@
+export { SelectContainer } from "./SelectContainer.tsx"
+export { SelectGroup } from "./SelectGroup.tsx"
+export { SelectValue } from "./SelectValue.tsx"
+export { SelectTrigger } from "./SelectTrigger.tsx"
+export { SelectContent } from "./SelectContent.tsx"
+export { SelectItem } from "./SelectItem.tsx"
