+++ src/shared/ui/table/index.ts
@@ -1,0 +1,6 @@
+export { Table } from "./Table.tsx"
+export { TableHead } from "./TableHead.tsx"
+export { TableHeader } from "./TableHeader.tsx"
+export { TableBody } from "./TableBody.tsx"
+export { TableRow } from "./TableRow.tsx"
+export { TableCell } from "./TableCell.tsx"
